updated init, created cluster.conf
updated neo4j_ini script and added cluser.config file

+++ scripts/Server/neo4j_init.sh
@@ -46,21 +46,12 @@ BACKUP_SCRIPT=neo4j_backup_routine.sh
 ADMIN_EMAIL=dongyang@microsoft.com
 # provide the crontab entry for the intrusion detection module
 AIDE_ENTRY="0 1 * * * /usr/sbin/aide --check"
-# detail the hostnames and ip addresses of all nodes in the cluster
-NODE1_HN=3DM-Graph01
-NODE2_HN=3DM-Graph02
-NODE3_HN=3DM-Graph03
-NODE1_IP=10.1.0.7
-NODE2_IP=10.1.0.6
-NODE3_IP=10.1.0.10
-# the node details of the script destination
-THISH=$NODE1_HN
-THISIP=$NODE1_IP
-THISNUM=1
+
+
 #------------------------------------------------
 # End script variables
 #------------------------------------------------
-if [ -e $NEOLOG]; then
+if [ -e $NEOLOG ]; then
 	echo "located log file"
 	echo "---------------------------------------------" >> $NOELOG
 	echo "$(date) : begin Neo4J installation logging --" >> $NEOLOG
@@ -69,16 +60,29 @@ else
 	echo "$(date) : begin Neo4J installation logging --" > $NEOLOG
 fi
 
+if [ -e $CLUSTER_CONFIG ]; then
+	echo "could not locate cluster configruation file, aborting script"
+	exit 1
+fi
+
+source $CLUSTER_CONFIG
+
+# the node details of the script destination
+THISH=$NODE1_HN
+THISIP=$NODE1_IP
+THISNUM=1
 
 #------------------------------------------------
 # Create Neo4j directory structures
 #------------------------------------------------
 
 echo "Creating Neo4j home and support directories"
-cd /
+
 mkdir -p $NEOETC && chmod 755 $NEOETC
 mkdir -p $NEOBIN && chmod 755 $NEOBIN
 mkdir -p $NEOBAK && chmod 755 $NEOBAK
+
+echo "$(date) neo4j directory structure created" >> $NEOLOG
 #
 logger -p local0.notice -t $LOGTAG "neo4j home and support directories created"
 
@@ -90,7 +94,7 @@ logger -p local0.notice -t $LOGTAG "neo4j home and support directories created"
 
 if [ ! -d /home/neo4j ]; then
 	echo "cannot locate neo4j home directory"
-	exit 1
+	exit 2
 fi
 
 #
@@ -129,12 +133,14 @@ ENDOC
 
 logger -p local0.notice -t $LOGTAG "user $USER profile updated"
 
+echo "neo4j user .bash_profile updated" >> $NEOLOG
+
 #------------------------------------------------
 # Populate Neo4j directories with Neo4j files
 #------------------------------------------------
 
 # return to the neo4j home directory
-
+CURRDIR=$(pwd)
 cd ~
 
 # execute the updated bash profile script
@@ -148,17 +154,29 @@ source .bash_profile
 
 # uncompress the neo4j support tarball
 
-tar -zxvf $SUPPORT_TGZ_FILE -C $NEOBASE && logger -p local0.notice -t $LOGTAG "support files loaded"
+echo "returning to original directory"
+
+cd $CURRDIR
+if [ -e $SUPPORT_TGZ_FILE ]; then
+	tar -zxvf $SUPPORT_TGZ_FILE -C $NEOBASE && logger -p local0.notice -t $LOGTAG "support files loaded"
+else
+	echo "unable to locate neo4j support tar file, aborting script, $SUPPORT_TGZ_FILE"
+	echo "$(date) ERROR:  unable to locate support tar file $SUPPORT_TGZ_FILE, aborting script"  >> $NEOLOG
+	exit 3
+fi
+
 
 # copying support files to proper directories
 
-cp $CONF_FILE $NEOETC/$CONF_FILE
-cp $DIAG_FILE $NEOBIN/$DIAG_FILE
-cp $SAMBA_FILE $NEOETC/$SAMBA_FILE
-cp $NEO4J_PROP_FILE $NEOETC/$NEO4J_PROP_FILE
-cp $NEO4J_SRV_PROP_FILE $NEOETC/$NEO4J_SRV_PROP_FILE
-cp $NEO4J_WRAP_FILE $NEOETC/$NEO4J_WRAP_FILE
-cp $MONITOR_SCRIPT $NEOETC/$MONITOR_SCRIPT
+echo "$(date) copying support files " >> $NEOLOG
+
+if [ -e $CONF_FILE ]; then cp $CONF_FILE $NEOETC/$CONF_FILE && echo "$CONF_FILE copied" ; fi
+if [ -e $DIAG_FILE ]; then cp $DIAG_FILE $NEOBIN/$DIAG_FILE && echo "$DIAG_FILE copied"; fi
+fi [ -e $SAMBA_FILE ]; then cp $SAMBA_FILE $NEOETC/$SAMBA_FILE && echo "$SAMBA_FILE copied"; fi
+if [ -e $NEO4J_PROP_FILE ]; then cp $NEO4J_PROP_FILE $NEOETC/$NEO4J_PROP_FILE && echo "$NEO$J_PROP_FILE copied"; fi
+if [ -e $NEO4J_SRV_PROP_FILE ]; then cp $NEO4J_SRV_PROP_FILE $NEOETC/$NEO4J_SRV_PROP_FILE && echo "$NEO4J_SRV_PROP_FILE copied"; fi
+if [ -e $NEO4J_WRAP_FILE ]; then cp $NEO4J_WRAP_FILE $NEOETC/$NEO4J_WRAP_FILE && echo "$NEO4J_WRAP_FILE copied"; fi
+if [ -e $MONITOR_SCRIPT ]; then cp $MONITOR_SCRIPT $NEOETC/$MONITOR_SCRIPT && echo "$MONITOR_SCRIPT copied"; fi
 
 logger -p local0.notice -t $LOGTAG "support files deployed"
 
@@ -178,74 +196,22 @@ if [ -d $NEOCONF ]; then
 	
 	# update server properties file with cluster configuration details
 	
-	sed -i 's/xthis_server_num/$THISNUM/g' $NEO4J_PROP_FILE
-	sed -i 's/xsrv1ip/$NODE1_IP/g' $NEO4J_PROP_FILE
-	sed -i 's/xsrv2ip/$NODE2_IP/g' $NEO4J_PROP_FILE
-	sed -i 's/xsrv3ip/$NODE3_IP/g' $NEO4J_PROP_FILE
-	sed -i 's/xthis_server_ip/$THISIP/g' $NEO4J_PROP_FILE
+	sed -i 's/xthis_server_num/$THISNUM/g' $NEO4J_PROP_FILE && echo "updated $NEO4J_PROP_FILE"
+	sed -i 's/xsrv1ip/$NODE1_IP/g' $NEO4J_PROP_FILE && echo "updated $NEO4J_PROP_FILE"
+	sed -i 's/xsrv2ip/$NODE2_IP/g' $NEO4J_PROP_FILE && echo "updated $NEO4J_PROP_FILE"
+	sed -i 's/xsrv3ip/$NODE3_IP/g' $NEO4J_PROP_FILE && echo "updated $NOE4J_PROP_FILE"
+	sed -i 's/xthis_server_ip/$THISIP/g' $NEO4J_PROP_FILE && echo "updated $NEO4J_PROP_FILE"
 	
 	
 	logger -p local0.notice -t $LOGTAG "neo4j cluster configuration updated"
 else
 	echo "cannot locate neo4j home directory"
 	echo "manual configuration of neo4j cluster is required"
+	echo "$(date) ERROR: unable to locate home directory" >> $NEOLOG
+	echo "$(date) ERROR: manual configuration of neo4j cluster required" >> $NEOLOG
 	
 	logger -p local0.notice -t $LOGTAG "unable to deploy neo4j ccluster configuration files"
 fi
-
-#------------------------------------------------
-# Install intrusion detection package
-#------------------------------------------------
-
-# check to see if the aide package is already installed
-
-AIDETEST=$(yum list installed aide |& grep Error | awk '{ print $1 }' | sed s/://) 
-
-# install the aide package if it is missing
-
-if [ $AIDETEST == "Error" ]
-then
-    echo "AIDE not installed, installing" && logger -p local0.notice -t $LOGTAG "installing AIDE"
-	yum install -y aide
-else
-    echo "AIDE already installed"
-fi
-
-# update the crontab mailto configuration with the admin email
-
-sed -i 's/root/$ADMIN_EMAIL/' /etc/crontab
-
-# initialize the aide package and get the service started
-
-aide --init
-
-logger -p local0.notice -t $LOGTAG "AIDE initialized"
-
-# navigate to the aide library directory
-# the aide service creates a new database
-# so we rename the new database to the default
-
-cd /var/lib/aide
-mv aide.db.new.gz aide.db.gz
-
-# according to the manual, we need to invoke the check & update routines
-# and then switch to the newly created database
-
-aide --check
-aide --update
-
-rm aide.db.gz
-
-mv aide.db.new.gz aide.db.gz
-
-# now we update the crontab with an entry for the aide package
-
-crontab -e
-
-cat $AIDE_ENTRY >> /etc/crontab
-
-logger -p local0.notice -t $LOGTAG "crontab updated with AIDE entry"
-
 
 #------------------------------------------------
 # Set Neo4j Server specific os configurations
@@ -255,23 +221,36 @@ logger -p local0.notice -t $LOGTAG "crontab updated with AIDE entry"
 # we need to update some server configuration files
 # in order for the neo4j cluster to function properly
 #
-
+if [ -e /etc/security/limits.conf ]; then
 cat << ENDOC >> /etc/security/limits.conf
 # NEO4J SUPPORT MODIFICATION
 neo4j   soft    nofile  40000
 neo4j   hard    nofile  40000
 # END NEO4J SUPPORT MODIFICATION
 ENDOC
-
+	
 logger -p local0.notice -t $LOGTAG "limits.conf modified"
+echo "modified limits.conf"
+else
+	echo "unable to locate limits.conf file"
+	echo "$(date) ERROR: unable to locate limits.conf file.  Manual config required" >> $NOELOG
+fi
 
+if [ -e /etc/pam.d/su ]; then
 cat << ENDOC >> /etc/pam.d/su
 # NEO4J SUPPORT MODIFICATION
 session    required   pam_limits.so
 # END NEO4J SUPPORT MODIFICATION
 ENDOC
+logger -p local0.notice -t $LOGTAG "/etc/pam.d/su file modified"
+echo "modified /etc/pam.d/su file"
 
-logger -p local0.notice -t $LOGTAG "pam.d/su modified"
+else
+	echo "unable to locate /etc/pam.d/su file"
+	echo "$(date) ERROR: unable to locate /etc/pam.d/su file.  Manual config required" >> $NEOLOG
+fi
+
+
 
 # A restart is required for the settings to take effect.
 # After the above procedure, the neo4j user will have a limit of 40 000 simultaneous open files.
@@ -282,30 +261,6 @@ echo "required server configuration changes complete."
 echo "System must be restarted before changes take affect"
 
 #------------------------------------------------
-# Install a jdk
-#------------------------------------------------
-
-# 
-# The neo4j cluser requires at minimum a java 6 jdk to be installed
-#
-
-# determine if the java 8 openjdk is installed.
-# We should improve this script by checking for multiple releases (6,7,8)
-# and multiple vendors (openjdk, Oracle).
-
-JDKTEST=$(yum list installed java-1.8.0-openjdk |& grep Error | awk '{ print $1 }' | sed s/://) 
-
-# if there is no jdk installed, install the openjdk 1.8.0 package
-
-if [ $JDKTEST == "Error" ]
-then
-    echo "JDK not installed, installing" && logger -p local0.notice -t $LOGTAG "installing JDK"
-	yum install -y java-1.8.0-openjdk
-else
-    echo "JDK already installed"
-fi
-
-#------------------------------------------------
 # Update network information
 #------------------------------------------------
 
@@ -314,7 +269,7 @@ fi
 # we add nodes to host file so we can reference
 # all nodes in the cluster by their actual hostnames
 #
-
+if [ -d /etc/hosts ]; then
 cat << ENDOC >> /etc/hosts
 # NEO4J SUPPORT MODIFICATION
 $NODE1_IP	$NODE1_HN
@@ -322,101 +277,12 @@ $NODE2_IP	$NODE2_HN
 $NODE3_IP	$NODE3_HN
 # END NEO4J SUPPORT MODIFICATION
 ENDOC
-
+echo "modified /etc/hosts file"
 logger -p local0.notice -t $LOGTAG "/etc/hosts modified"
-
-#------------------------------------------------
-# Install and configure auditing
-#------------------------------------------------
-
-#
-# To ensure we meet the standard security certification guidelines,
-# we need to make sure auditing is installed and running rules
-# based on the considered requirements (stig, capp, nipsom, etc...).
-# 
-
-# check to see if auditing is installed and install if it is missing
-
-AUDITTEST=$(yum list installed audit |& grep Error | awk '{ print $1 }' | sed s/://) 
-
-if [ $AUDITTEST == "Error" ]
-then
-    echo "AUDIT not installed, installing" && logger -p local0.notice -t $LOGTAG "installing AUDIT"
-	yum install -y audit
 else
-    echo "audit already installed"
+	echo "unable to locate /etc/hosts file"
+	echo "$(date) ERROR: unable to locate /etc/hosts file. Manual config required"
 fi
-
-# set audit rules -- here I use stig because it's thorough
-
-
-
-auditctl -R /usr/share/doc/audit-version/stig.rules
-if [ $? -eq 0 ]; then
-	echo "audit rules loaded"
-	logger -p local0.notice -t $LOGTAG "auditing rules loaded and auditing started"
-else
-	echo "unable to reload audit daemon with new rules"
-	echo "please execute auditctl -R and provide valid rules file"
-	logger -p local0.notice -t $LOGTAG "erro loading audit rules"
-fi
-
-
-# optionally you can copy the specific rules file to the default:
-#
-# cp /etc/audit/audit.rules /etc/audit/audit.rules_backup
-# cp /usr/share/doc/audit-version/stig.rules /etc/audit/audit.rules
-#
-# or nipsom.rules
-# or capp.rules
-# or lspp.rules
-
-#------------------------------------------------
-# Setup and configure samba
-#------------------------------------------------
-
-#
-# Next we need to setup samba so we can load files from windows
-# without the need for ssh or ftp.
-#
-
-# call samba config script
-
-cd $NEOBIN
-chmod +x $SAMBA_SCRIPT
-./$SAMBA_SCRIPT
-RETVAL=$?
-if [ $RETVAL -ne 0 ]; then
-	echo "samba script failed at exit $RETVAL"
-	exit 1
-else
-	echo "samba script successful"
-fi
-
-
-#------------------------------------------------
-# Configure IPTABLES
-#------------------------------------------------
-
-#
-# Although the Azure virtual switches control most of our network security
-# we want to ensure that no local firewall rules prevent
-# our neo4j cluser or other services (i.e. samba) from operating normally
-#
-
-# call the firewall configuration script
-
-cd $NEOBIN
-chmod +x $FIREWALL_SCRIPT
-./FIREWALL_SCRIPT
-RETVAL=$?
-if [ $RETVAL -ne 0 ]; then
-	echo "firewall script failed at exit $RETVAL"
-	exit 1
-else
-	echo "firewall script successful"
-fi
-
 
 #------------------------------------------------
 # Configure backup scheduler
@@ -430,9 +296,14 @@ fi
 # but will not be copied to /etc/cron.daily
 #
 
-cp $BACKUP_SCRIPT $NEOBIN/$BACKUP_SCRIPT
-echo "backup script copied to $NEOBIN"
-echo "to schedule backup routine copy  $BACKUP_SCRIPT to /etc/cron.daily"
+if [ -e $BACKUP_SCRIPT ]; then
+	cp $BACKUP_SCRIPT $NEOBIN/$BACKUP_SCRIPT
+	echo "backup script copied to $NEOBIN"
+	echo "to schedule backup routine copy  $BACKUP_SCRIPT to /etc/cron.daily"
+else
+	echo "could not locate backup routine script"
+	echo "$(date) ERROR:  could not locate backup routine script"
+fi
 
 #------------------------------------------------
 # Deploy monitoring script
@@ -444,9 +315,15 @@ echo "to schedule backup routine copy  $BACKUP_SCRIPT to /etc/cron.daily"
 # Other monitoring routines should be added to the script.
 # 
 
-cp $MONITOR_SCRIPT /etc/cron.daily/$MONITOR_SCRIPT && chmod +x /etc/cron.daily/$MONITOR_SCRIPT
+if [ -e $MONITOR_SCRIPT ]; then
+	cp $MONITOR_SCRIPT /etc/cron.daily/$MONITOR_SCRIPT && chmod +x /etc/cron.daily/$MONITOR_SCRIPT
+	echo "copied monitor script to /etc/cron.daily/"
+	logger -p local0.notice -t $LOGTAG "file : $MONITOR_SCRIPT added to cron.daily"
+else
+	echo "could not locate monitor script $MONITOR_SCRIPT"
+	echo "$(date) ERROR:  could not locate monitor script, $MONITOR_SCRIPT"
+fi
 
-logger -p local0.notice -t $LOGTAG "file : $MONITOR_SCRIPT added to cron.daily"
 
 
 #------------------------------------------------
@@ -459,7 +336,13 @@ logger -p local0.notice -t $LOGTAG "file : $MONITOR_SCRIPT added to cron.daily"
 # of diagnostic information for use with troubleshooting production issues.
 #
 
-cp $SUPPORT_SCRIPT $NEOBIN/$SUPPORT_SCRIPT && chmod +x $NEOBIN/$SUPPORT_SCRIPT
+if [ -e $SUPPORT_SCRIPT ]; then
+	cp $SUPPORT_SCRIPT $NEOBIN/$SUPPORT_SCRIPT && chmod +x $NEOBIN/$SUPPORT_SCRIPT
+	echo "copied support script"
+else	
+	echo "could not locate support script, $SUPPORT_SCRIPT"
+	echo "$(date) ERROR: could not locate support script, $SUPPORT_SCRIPT"
+fi
 
 #------------------------------------------------
 # Ensure proper ownership of directories and files
@@ -474,6 +357,9 @@ chown -R neo4j:neo4j $NEOBASE
 chown -R neo4j:neo4j $NEOSUPP
 
 logger -p local0.notice -t $LOGTAG "neo4j file and directory ownership changed"
+
+echo "$(date) Neo4j_init.sh script fiinished"
+
